+++ src/services/AegisAIService.ts
@@ -1,3 +1,5 @@
+import { type Ref, ref } from 'vue';
+
 import type {
   AegisAICVEAnalysisParamsType,
   AegisAIComponentAnalysisParamsType,
@@ -132,6 +134,14 @@ function queryStringFromParams(params: Record<string, any>) {
 }
 
 export class AegisAIService {
+  isFetching: Ref<boolean>;
+  requestDuration: Ref<number>;
+
+  constructor() {
+    this.isFetching = ref(false);
+    this.requestDuration = ref(0);
+  }
+
   /**
    * Analyze a component using AEGIS-AI intelligence
    * @param params Component analysis parameters
@@ -139,7 +149,7 @@ export class AegisAIService {
    */
   async analyzeComponent(params: AegisAIComponentAnalysisParamsType): Promise<any> {
     try {
-      const result = await aegisAIFetch({
+      const result = await this.fetch({
         method: 'GET',
         url: '/analysis/component',
         params: {
@@ -163,7 +173,7 @@ export class AegisAIService {
    */
   async analyzeCVE(params: AegisAICVEAnalysisParamsType): Promise<any> {
     try {
-      const result = await aegisAIFetch({
+      const result = await this.fetch({
         method: 'GET',
         url: '/analysis/cve',
         params: {
@@ -189,12 +199,13 @@ export class AegisAIService {
   async analyzeCVEWithContext(
     params: { feature: 'suggest-cwe' } & AegisAICVEAnalysisWithContextParamsType,
   ): Promise<CweSuggestionDetails>;
+
   async analyzeCVEWithContext(params: AegisAICVEAnalysisWithContextParamsType): Promise<any>;
   async analyzeCVEWithContext(params: AegisAICVEAnalysisWithContextParamsType): Promise<any> {
     try {
       const { detail, feature, ...contextData } = params;
 
-      const result = await aegisAIFetch({
+      const result = await this.fetch({
         method: 'POST',
         url: `/analysis/cve/${feature}`,
         params: {
@@ -210,6 +221,17 @@ export class AegisAIService {
     }
   }
 
+  async fetch(config: AegisAIFetchOptions, factoryOptions?: AegisAIFetchCallbacks): ReturnType<typeof aegisAIFetch> {
+    this.isFetching.value = true;
+    const requestStartTime = Date.now();
+    try {
+      return await aegisAIFetch(config, factoryOptions);
+    } finally {
+      this.isFetching.value = false;
+      this.requestDuration.value = Date.now() - requestStartTime;
+    }
+  }
+
   /**
    * Generate a response using AEGIS-AI (for the web console)
    * @param prompt The prompt to send to AEGIS-AI
@@ -217,7 +239,7 @@ export class AegisAIService {
    */
   async generateResponse(prompt: string): Promise<any> {
     try {
-      const result = await aegisAIFetch({
+      const result = await this.fetch({
         method: 'POST',
         url: '/generate_response',
         data: { prompt },
@@ -235,7 +257,7 @@ export class AegisAIService {
    */
   async getConsole(): Promise<string> {
     try {
-      const result = await aegisAIFetch({
+      const result = await this.fetch({
         method: 'GET',
         url: '/console',
       });