+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 ### Added
 * Allow grouping affects by ps_module (`OSIDB-4546`)
 
+## [2025.11.0]
+### Added
+* Tour guide for new features (`OSIDB-4583`)
+* Allow grouping affects by ps_module (`OSIDB-4546`)
+
+
 ## [2025.10.0]
 ### Fixed
 * Fix multi-row operations in affects table (`OSIDB-4381`)
@@ -526,7 +532,8 @@ The first release for user testing, briefly reaching feature parity with OSIDB
 ### Added
 * Early repo layout & Flaw template
 
-[Unreleased]: https://github.com/RedHatProductSecurity/osim/compare/v2025.10.0...HEAD
+[Unreleased]: https://github.com/RedHatProductSecurity/osim/compare/v2025.11.0...HEAD
+[2025.11.0]: https://github.com/RedHatProductSecurity/osim/compare/v2025.10.0...v2025.11.0
 [2025.10.0]: https://github.com/RedHatProductSecurity/osim/compare/v2025.9.0...v2025.10.0
 [2025.9.0]: https://github.com/RedHatProductSecurity/osim/compare/v2025.8.0...v2025.9.0
 [2025.8.0]: https://github.com/RedHatProductSecurity/osim/compare/v2025.7.0...v2025.8.0